refactor types and build keyv
