+++ package.json
@@ -10,7 +10,8 @@
     "packages/*"
   ],
   "scripts": {
-    "build": "yarn workspaces run build",
+    "build:keyv": "cd packages/keyv && yarn build",
+    "build": "yarn build:keyv && yarn workspaces run build",
     "test": " yarn build && c8 --reporter=lcov yarn workspaces run test:ci",
     "test:services:start": "docker-compose -f ./docker-compose.yaml up -d",
     "test:services:stop": "docker-compose -f ./docker-compose.yaml down -v",
